Deleting selected text between cells deletes just the text.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@8541 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/ptbl/xp/pt_PT_DeleteSpan.cpp
@@ -49,7 +49,8 @@
 /****************************************************************/
 bool pt_PieceTable::deleteSpan(PT_DocPosition dpos1,
 							   PT_DocPosition dpos2,
-							   PP_AttrProp *p_AttrProp_Before, bool bDontGlob)
+							   PP_AttrProp *p_AttrProp_Before, 
+							   bool bDontGlob)
 {
 	if(m_pDocument->isMarkRevisions())
 	{
@@ -104,6 +105,10 @@ bool pt_PieceTable::deleteSpan(PT_DocPosition dpos1,
 					case PTX_Section:
 					case PTX_SectionHdrFtr:
 					case PTX_SectionEndnote:
+					case PTX_SectionTable:
+					case PTX_SectionCell:
+					case PTX_EndTable:
+					case PTX_EndCell:
 						iLen = pf_FRAG_STRUX_SECTION_LENGTH;
 						break;
 
@@ -148,7 +153,7 @@ bool pt_PieceTable::deleteSpan(PT_DocPosition dpos1,
 				if(   (pRev->getType() == PP_REVISION_ADDITION)
 				   || (pRev->getType() == PP_REVISION_ADDITION_AND_FMT ))
 				{
-					if(!_realDeleteSpan(dpos1, dposEnd, p_AttrProp_Before, bDontGlob))
+					if(!_realDeleteSpan(dpos1, dposEnd, p_AttrProp_Before,true, bDontGlob))
 						return false;
 
 					dpos1 = dposEnd;
@@ -188,7 +193,160 @@ bool pt_PieceTable::deleteSpan(PT_DocPosition dpos1,
 		return true;
 	}
 	else
-		return _realDeleteSpan(dpos1, dpos2, p_AttrProp_Before, bDontGlob);
+		return _realDeleteSpan(dpos1, dpos2, p_AttrProp_Before, true, bDontGlob);
+}
+
+/****************************************************************/
+bool pt_PieceTable::deleteSpanWithTable(PT_DocPosition dpos1,
+							   PT_DocPosition dpos2,
+							   PP_AttrProp *p_AttrProp_Before, 
+							   bool bDeleteTableStruxes)
+{
+	bool bDontGlob = false;
+	if(m_pDocument->isMarkRevisions())
+	{
+		const XML_Char name[] = "revision";
+		const XML_Char * pRevision = NULL;
+
+		// first retrive the starting and ending fragments
+		pf_Frag * pf1, * pf2;
+		PT_BlockOffset Offset1, Offset2;
+
+		if(!getFragsFromPositions(dpos1,dpos2, &pf1, &Offset1, &pf2, &Offset2))
+			return false;
+
+		// now we have to traverse the fragments and change their
+		// formatting
+
+		pf_Frag * pTemp;
+		pf_Frag * pEnd = pf2->getNext();
+		pf_Frag * pNext;
+
+
+		for(pTemp = pf1; pTemp != pEnd; pTemp = pNext)
+		{
+			// we cannot ask for the next in the for statement,
+			// because we might have deleted that fragment by then
+			pNext = pTemp->getNext();
+
+			// get attributes for this fragement
+			const PP_AttrProp * pAP;
+			pf_Frag::PFType eType = pTemp->getType();
+			UT_uint32 iLen;
+			PTStruxType eStruxType;
+
+			if(eType == pf_Frag::PFT_Text)
+			{
+				if(!getAttrProp(((pf_Frag_Text*)pTemp)->getIndexAP(),&pAP))
+					return false;
+			}
+			else if(eType == pf_Frag::PFT_Strux)
+			{
+				if(!getAttrProp(((pf_Frag_Strux*)pTemp)->getIndexAP(),&pAP))
+					return false;
+
+				eStruxType = ((pf_Frag_Strux*)pTemp)->getStruxType();
+
+				switch (eStruxType)
+				{
+					case PTX_Block:
+						iLen = pf_FRAG_STRUX_BLOCK_LENGTH;
+						break;
+
+					case PTX_Section:
+					case PTX_SectionHdrFtr:
+					case PTX_SectionEndnote:
+					case PTX_SectionTable:
+					case PTX_SectionCell:
+					case PTX_EndCell:
+					case PTX_EndTable:
+						iLen = pf_FRAG_STRUX_SECTION_LENGTH;
+						break;
+
+					default:
+						UT_ASSERT(UT_NOT_IMPLEMENTED);
+						iLen = 1;
+						break;
+				}
+
+			}
+			else if(eType == pf_Frag::PFT_Object)
+			{
+				if(!getAttrProp(((pf_Frag_Object*)pTemp)->getIndexAP(),&pAP))
+					return false;
+			}
+			else
+			{
+				// something that does not carry AP
+				continue;
+			}
+
+			if(!pAP->getAttribute(name, pRevision))
+				pRevision = NULL;
+
+			PP_RevisionAttr Revisions(pRevision);
+
+			// now we need to see if revision with this id is already
+			// present, and if it is, whether it might not be addition
+			UT_uint32 iId = m_pDocument->getRevisionId();
+			const PP_Revision * pRev = Revisions.getGreatestLesserOrEqualRevision(iId);
+
+			PT_DocPosition dposEnd = UT_MIN(dpos2,dpos1 + pTemp->getLength());
+
+			if(pRev && iId == pRev->getId())
+			{
+				// OK, we already have a revision with this id here,
+				// which means that the editor made a change earlier
+				// (insertion or format change) but now wants this deleted
+				//
+				// so if the previous revision is an addition, we just
+				// remove this fragment as if this was regular delete
+				if(   (pRev->getType() == PP_REVISION_ADDITION)
+				   || (pRev->getType() == PP_REVISION_ADDITION_AND_FMT ))
+				{
+					if(!_realDeleteSpan(dpos1, dposEnd, p_AttrProp_Before,bDeleteTableStruxes,
+										bDontGlob))
+						return false;
+
+					dpos1 = dposEnd;
+					continue;
+				}
+			}
+
+			Revisions.addRevision(iId,PP_REVISION_DELETION,NULL,NULL);
+			const XML_Char * ppRevAttrib[3];
+			ppRevAttrib[0] = name;
+			ppRevAttrib[1] = Revisions.getXMLstring();
+			ppRevAttrib[2] = NULL;
+
+			switch (eType)
+			{
+				case pf_Frag::PFT_Text:
+					if(! _realChangeSpanFmt(PTC_AddFmt, dpos1, dposEnd, ppRevAttrib,NULL))
+						return false;
+					break;
+
+				case pf_Frag::PFT_Strux:
+					if(! _realChangeStruxFmt(PTC_AddFmt, dpos1 + iLen, dpos1 + 2*iLen, ppRevAttrib,NULL,eStruxType))
+						return false;
+					break;
+#if 0
+				case pf_Frag::PFT_Object:
+					if(! _realChangeStruxFmt(PTC_AddFmt, dpos1, dposEnd, ppRevAttrib,NULL))
+						return false;
+					break;
+#endif
+				default:;
+			}
+
+			dpos1 = dposEnd;
+		}
+
+		return true;
+	}
+	else
+		return _realDeleteSpan(dpos1, dpos2, p_AttrProp_Before, bDeleteTableStruxes,
+							   bDontGlob);
 }
 
 
@@ -630,12 +788,21 @@ bool pt_PieceTable::_deleteFormatting(PT_DocPosition dpos1,
 	return true;
 }
 
+bool pt_PieceTable::_StruxIsNotTable(pf_Frag_Strux * pfs)
+{
+	PTStruxType its = pfs->getStruxType();
+	bool b = ((its != PTX_SectionTable) && (its != PTX_SectionCell)
+			  && (its != PTX_EndTable) && (its != PTX_EndCell));
+	return b;
+}
+
 bool pt_PieceTable::_deleteComplexSpan(PT_DocPosition dpos1,
-									   PT_DocPosition dpos2)
+									   PT_DocPosition dpos2,
+									   bool bDeleteTableStruxes)
 {
 	pf_Frag * pfNewEnd;
 	UT_uint32 fragOffsetNewEnd;
-
+	bool bPrevWasCell = false;
 	pf_Frag * pf_First;
 	pf_Frag * pf_End;
 	PT_BlockOffset fragOffset_First;
@@ -667,10 +834,40 @@ bool pt_PieceTable::_deleteComplexSpan(PT_DocPosition dpos1,
 
 		case pf_Frag::PFT_Strux:
 		{
+//
+// OK this code is leave the cell/table sctructures in place unless we defiantely want to delete
+// them.
+// 
 			pf_Frag_Strux * pfs = static_cast<pf_Frag_Strux *> (pf_First);
-
-			bool bResult = _deleteStruxWithNotify(dpos1,pfs,
+			bool bResult = true;
+			if(bDeleteTableStruxes || _StruxIsNotTable(pfs))
+			{
+				if(bPrevWasCell && pfs->getStruxType()== PTX_Block)
+				{
+					bPrevWasCell = false;
+					pfNewEnd = pfs->getNext();
+					fragOffsetNewEnd = 0;
+					pfsContainer = pfs;
+				}
+				else
+				{
+					bResult = _deleteStruxWithNotify(dpos1,pfs,
 												  &pfNewEnd,&fragOffsetNewEnd);
+				}
+			}
+			else
+			{
+				if(pfs->getStruxType() == PTX_SectionCell)
+				{
+					bPrevWasCell = true;
+				}
+				else
+				{
+					bPrevWasCell = false;
+				}
+				pfNewEnd = pfs->getNext();
+				fragOffsetNewEnd = 0;
+			}
 			UT_ASSERT(bResult);
 			// we do not update pfsContainer because we just deleted pfs.
 		}
@@ -1337,8 +1534,10 @@ bool pt_PieceTable::_deleteComplexSpan_norec(PT_DocPosition dpos1,
 
 
 bool pt_PieceTable::_realDeleteSpan(PT_DocPosition dpos1,
-							   PT_DocPosition dpos2,
-							   PP_AttrProp *p_AttrProp_Before, bool bDontGlob)
+									PT_DocPosition dpos2,
+									PP_AttrProp *p_AttrProp_Before,
+									bool bDeleteTableStruxes, 
+									bool bDontGlob)
 {
 	// remove (dpos2-dpos1) characters from the document at the given position.
 
@@ -1394,7 +1593,7 @@ bool pt_PieceTable::_realDeleteSpan(PT_DocPosition dpos1,
 		//  If the delete is sure to be within a fragment, we don't
 		//  need to worry about much of the bookkeeping of a complex
 		//  delete.
-		bSuccess = _deleteComplexSpan(dpos1, dpos2);
+		bSuccess = _deleteComplexSpan(dpos1, dpos2,bDeleteTableStruxes);
 	}
 	else
 	{
@@ -1405,7 +1604,7 @@ bool pt_PieceTable::_realDeleteSpan(PT_DocPosition dpos1,
 
 		bSuccess = _deleteFormatting(dpos1, dpos2);
 		if (bSuccess)
-			bSuccess = _deleteComplexSpan(dpos1, dpos2);
+			bSuccess = _deleteComplexSpan(dpos1, dpos2,bDeleteTableStruxes);
 
 		while (bSuccess && stDelayStruxDelete.getDepth() > 0)
 		{
@@ -1415,8 +1614,17 @@ bool pt_PieceTable::_realDeleteSpan(PT_DocPosition dpos1,
 
  			pf_Frag *pf;
 			PT_DocPosition dp;
-			bSuccess = _deleteStruxWithNotify(dpos1 - pfs->getLength(), pfs,
+			if(bDeleteTableStruxes || _StruxIsNotTable(pfs))
+			{
+				bSuccess = _deleteStruxWithNotify(dpos1 - pfs->getLength(), pfs,
 											  &pf, &dp);
+			}
+			else
+			{
+				bSuccess = true;
+				pf = pfs->getNext();
+				dp = 0;
+			}
 		}
 
 		_changePointWithNotify(dpos1);
